% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\docType{data}
\name{APPL}
\alias{APPL}
\title{InterProScan applications}
\format{
An object of class \code{character} of length 24.
}
\usage{
APPL
}
\description{
InterProScan applications
}
\details{
All supported application values for InterProScan5
}
\keyword{datasets}
