% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_time_hist.R
\name{job_time_hist}
\alias{job_time_hist}
\title{Plot a histogram of InterProScan API runtimes submitted with iprscanR}
\usage{
job_time_hist(ipr_log_path)
}
\arguments{
\item{ipr_log_path}{path to the iprscan api log}
}
\value{
A histogram plot object
}
\description{
Plot a histogram of InterProScan API runtimes submitted with iprscanR
}
