% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\docType{data}
\name{SEQTYPE}
\alias{SEQTYPE}
\title{InterProScan5 sequence type values}
\format{
An object of class \code{character} of length 2.
}
\usage{
SEQTYPE
}
\description{
InterProScan5 sequence type values
}
\details{
InterProScan5 can accept Amino Acid (protein) or Nucleic Acid (DNA) sequences. The API values are 'p' and 'n', respectively
}
\keyword{datasets}
