% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_ipr.R
\name{submit_ipr}
\alias{submit_ipr}
\title{Submit IPRscan analysis for protein sequences (multifasta input)}
\usage{
submit_ipr(path2seq, outfolder, user_email)
}
\arguments{
\item{path2seq}{path to your sequence file}

\item{outfolder}{location for IPRscan outputs}

\item{email}{required email ID for job submission (to track and prevent accidental misuse of the API)}
}
\description{
Submit IPRscan analysis for protein sequences (multifasta input)
}
\keyword{architectures,}
\keyword{characterization}
\keyword{domain}
\keyword{domains,}
\keyword{protein}
