% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_ipr.R
\name{submit_ipr}
\alias{submit_ipr}
\title{Submit sequences for InterProScan5 analysis}
\usage{
submit_ipr(
  path2seq,
  outfolder,
  email,
  appl = c("PfamA"),
  stype = "p",
  max_polls = 300
)
}
\arguments{
\item{path2seq}{\itemize{
\item \strong{str} path to input FASTA file
}}

\item{outfolder}{\itemize{
\item \strong{str} location for IPRscan output directory
}}

\item{email}{\itemize{
\item \strong{str} required email ID for job submission
(to track and prevent accidental misuse of the API)
}}

\item{appl}{\itemize{
\item \strong{chr vec} a vector containing at least one of the annotation
}}

\item{stype}{\itemize{
\item \strong{str} "Protein" OR "Nucleic Acid"
methods provided by \href{https://www.ebi.ac.uk/Tools/services/rest/iprscan5/parameterdetails/appl}{InterProScan}
}}
}
\description{
Submit sequences for InterProScan5 analysis
}
\keyword{architectures,}
\keyword{characterization}
\keyword{domain}
\keyword{domains,}
\keyword{protein}
